+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+## Release 1.5.1a0 (WIP)
+
+### Breaking Changes
+- Change default value for number of hyperparameter optimization trials from 10 to 500. (@ernestum)
+- Derive number of intermediate pruning evaluations from number of time steps (1 evaluation per 100k time steps.) (@ernestum)
+
+### New Features
+- Support setting PyTorch's device with thye `--device` flag (@gregwar)
+
+### Bug fixes
+
+### Documentation
+
+### Other
+
+## Release 1.5.0 (2022-03-25)
+
+**Support for Weight and Biases experiment tracking**
+
+### Breaking Changes
+- Upgrade to Stable-Baselines3 (SB3) >= 1.5.0
+- Upgrade to sb3-contrib >= 1.5.0
+- Upgraded to gym 0.21
+
+### New Features
+- Verbose mode for each trial (when doing hyperparam optimization) can now be activated using the debug mode (verbose == 2)
+- Support experiment tracking via Weights and Biases via the `--track` flag (@vwxyzjn)
+- Support tracking raw episodic stats via `RawStatisticsCallback` (@vwxyzjn, see https://github.com/DLR-RM/rl-baselines3-zoo/pull/216)
+
+### Bug fixes
+- Policies saved during during optimization with distributed Optuna load on new systems (@jkterry)
+- Fixed script for recording video that was not up to date with the enjoy script
+
+### Documentation
+
+### Other
+
 ## Release 1.4.0 (2022-01-19)
 
 ### Breaking Changes
@@ -72,7 +109,7 @@
 - `HER` is now a replay buffer class and no more an algorithm
 - Removed `PlotNoiseRatioCallback`
 - Removed `PlotActionWrapper`
-- Changed `'lr'` key in Optuna param dict to `'learning_rate'` so the dict can be directly passed to SB3 methods (@justinkterry)
+- Changed `'lr'` key in Optuna param dict to `'learning_rate'` so the dict can be directly passed to SB3 methods (@jkterry)
 
 ### New Features
 - Add support for recording videos of best models and checkpoints (@mcres)
@@ -83,7 +120,7 @@
 - Added `--load-last-checkpoint` option for the enjoy script
 - Save Optuna study object at the end of hyperparameter optimization and plot the results (`plotly` package required)
 - Allow to pass multiple folders to `scripts/plot_train.py`
-- Flag to save logs and optimal policies from each training run (@justinkterry)
+- Flag to save logs and optimal policies from each training run (@jkterry)
 
 ### Bug fixes
 - Fixed video rendering for PyBullet envs on Linux
@@ -102,9 +139,9 @@
 - Minimum cloudpickle version added to `requirements.txt` (@amy12xx)
 - Fixed atari-py version (ROM missing in newest release)
 - Updated `SAC` and `TD3` search spaces
-- Cleanup eval_freq documentation and variable name changes (@justinkterry)
-- Add clarifying print statement when printing saved hyperparameters during optimization (@justinkterry)
-- Clarify n_evaluations help text (@justinkterry)
+- Cleanup eval_freq documentation and variable name changes (@jkterry)
+- Add clarifying print statement when printing saved hyperparameters during optimization (@jkterry)
+- Clarify n_evaluations help text (@jkterry)
 - Simplified hyperparameters files making use of defaults
 - Added new TQC+HER agents
 - Add `panda-gym`environments (@qgallouedec)